Move common station data into base file
